APIs/redhat.com/catalog_inventory/1.0.0/python-aiohttp-client/tests/test_service_inventory_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web

from openapi_server.models.error_not_found import ErrorNotFound
from openapi_server.models.service_inventories_collection import ServiceInventoriesCollection
from openapi_server.models.service_inventory import ServiceInventory
from openapi_server.models.tag import Tag
from openapi_server.models.tags_collection import TagsCollection


pytestmark = pytest.mark.asyncio

async def test_list_service_inventories(client):
    """Test case for list_service_inventories

    List ServiceInventories
    """
    params = [('limit', 100),
                    ('offset', 0),
                    ('filter', None),
                    ('sort_by', None)]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    response = await client.request(
        method='GET',
        path='//api/catalog-inventory/v1.0/service_inventories',
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_list_service_inventory_tags(client):
    """Test case for list_service_inventory_tags

    List Tags for ServiceInventory
    """
    params = [('limit', 100),
                    ('offset', 0),
                    ('filter', None),
                    ('sort_by', None)]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    response = await client.request(
        method='GET',
        path='//api/catalog-inventory/v1.0/service_inventories/{id}/tags'.format(id='id_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_show_service_inventory(client):
    """Test case for show_service_inventory

    Show an existing ServiceInventory
    """
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    response = await client.request(
        method='GET',
        path='//api/catalog-inventory/v1.0/service_inventories/{id}'.format(id='id_example'),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_tag_service_inventory(client):
    """Test case for tag_service_inventory

    Tag a ServiceInventory
    """
    body = {"tag":"/namespace/architecture=x86_64"}
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    response = await client.request(
        method='POST',
        path='//api/catalog-inventory/v1.0/service_inventories/{id}/tag'.format(id='id_example'),
        headers=headers,
        json=body,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_untag_service_inventory(client):
    """Test case for untag_service_inventory

    Untag a ServiceInventory
    """
    body = {"tag":"/namespace/architecture=x86_64"}
    headers = { 
        'Content-Type': 'application/json',
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    response = await client.request(
        method='POST',
        path='//api/catalog-inventory/v1.0/service_inventories/{id}/untag'.format(id='id_example'),
        headers=headers,
        json=body,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

